+++ joulescope/parameter.py
@@ -17,42 +17,52 @@ class Parameter:
     """Define a single parameter accessible by the application.
 
     :param name: The parameter name.
-    :param permission: The parameter permission which is either
-        'r' (read-only) or 'rw' (read-write).
     :param path: None if the parameter is only changed at the streaming start.
         'setting' if the parameter is part of the settings packet.
     :param default: The default value.
-    :param values: The list of acceptable values as ('name', value, ['alias1', ...]) tuples.
+    :param options: The list of acceptable values as ('name', value, ['alias1', ...]) tuples.
     :param units: The units string for the parameter.
+    :param brief: The brief user-meaningful description.
+    :param detail: The detailed user-meaningful description.
+    :param flags: The list of flags which include:
+        * developer: This parameter is intended for developers only, not end users.
+        * read_only: This parameter can only be read (no write permission).
+    :param validator: Function called to validate the parameter value.
     """
-    def __init__(self, name, permission, path, default, values, units=None):
+    def __init__(self, name, path, default=None, options=None, units=None,
+                 brief=None, detail=None,
+                 flags=None, validator=None):
         self.name = name
-        self.permission = permission
         self.default = default
         self.path = path
-        self.values = []
+        self.options = []
         self.str_to_value = {}
         self.units = units
-        if permission not in ['rw', 'r']:
-            raise ValueError('Parameter %s, invalid permission %r' % (name, permission))
-        if values is None:
+        self.brief = brief
+        self.detail = detail
+        self.flags = [] if flags is None else flags
+        self.validator = validator
+        if options is None:
             return
-        for idx, value in enumerate(values):
-            if len(value) < 2 or len(value) > 3:
-                raise ValueError('Parameter %s, value %d invalid: %r' % (name, idx, value))
-            vname = value[0]
-            vvalue = value[1]
+        for idx, option in enumerate(options):
+            if len(option) < 2 or len(option) > 3:
+                raise ValueError('Parameter %s, value %d invalid: %r' % (name, idx, option))
+            vname = option[0]
+            vvalue = option[1]
             if not isinstance(vname, str):
                 raise ValueError('Parameter %s, value %d invalid name %r' % (name, idx, vname))
-            if len(value) == 2:
-                value = (vname, vvalue, [])
-            self.values.append(value)
+            if len(option) == 2:
+                option = (vname, vvalue, [])
+            self.options.append(option)
             self._insert(vname, vvalue)
-            for alias in value[2]:
+            for alias in option[2]:
                 self._insert(alias, vvalue)
 
-        if default not in [x[0] for x in values]:
-            raise ValueError('Parameter %s, default %r not in ' % (name, permission))
+        if default not in [x[0] for x in options]:
+            raise ValueError('Parameter %s, default %r not found' % (name, default))
+
+    def __str__(self):
+        return f'Parameter({self.name})'
 
     def _insert(self, key, value):
         if key in self.str_to_value: